ajout d'un if pour gérer le choix de l'utilisateur

+++ scanner.sh
@@ -57,34 +57,38 @@ gerer_anomalie() {
 
     echo "Entrez votre choix (1/2/3) : "
     read choix
-    echo "Vous avez choisi : $choix"
 
-    case "$choix" in
-        1)
-            echo "Tentative de tuer le processus $pid..."
-            echo "Sauvegarde des informations du processus $pid avant de le tuer..."
-            ps -p "$pid" -o pid,comm,user,%mem,%cpu,state >> /var/log/process_suspects.log
-            if kill -0 "$pid" 2>/dev/null; then
-                kill -9 "$pid"
-                if [[ ! -e /var/log/process_monitor.log ]]; then
-                    touch /var/log/process_monitor.log
+    if (($choix!=null || $choix!=""));then
+        echo "Vous avez choisi : $choix"
+
+        case "$choix" in
+            1)
+                echo "Tentative de tuer le processus $pid..."
+                echo "Sauvegarde des informations du processus $pid avant de le tuer..."
+                ps -p "$pid" -o pid,comm,user,%mem,%cpu,state >> /var/log/process_suspects.log
+                if kill -0 "$pid" 2>/dev/null; then
+                    kill -9 "$pid"
+                    if [[ ! -e /var/log/process_monitor.log ]]; then
+                        touch /var/log/process_monitor.log
+                    fi
+                    echo "Processus $pid tué." >> /var/log/process_monitor.log
+                else
+                    echo "Le processus $pid n'existe plus." >> /var/log/process_monitor.log
                 fi
-                echo "Processus $pid tué." >> /var/log/process_monitor.log
-            else
-                echo "Le processus $pid n'existe plus." >> /var/log/process_monitor.log
-            fi
-            ;;
-        2)
-            renice 10 "$pid"
-            echo "Priorité du processus $pid baissée." >> journalEtatProcessus.txt
-            ;;
-        3)
-            echo "Processus $pid ignoré." >> journalEtatProcessus.txt
-            ;;
-        *)
-            echo "Choix invalide." >> journalEtatProcessus.txt
-            ;;
-    esac    
+                ;;
+            2)
+                renice 10 "$pid"
+                echo "Priorité du processus $pid baissée." >> journalEtatProcessus.txt
+                ;;
+            3)
+                echo "Processus $pid ignoré." >> journalEtatProcessus.txt
+                ;;
+            *)
+                echo "Choix invalide." >> journalEtatProcessus.txt
+                ;;
+        esac    
+    fi
+
 }
 
 enregistrer_etat_processus &
